+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix display of match path in results that match a document or corpus (metadata
   search) and the corpus has a special character like an umlaut.
+- Do not overwrite existing `service.toml` file (which might get the file
+  corrupted in certain cases) but always write the updated configuration to a
+  temporary TOML-file.
 
 ## [4.9.7] - 2022-09-22